+++ R/as.data.frame.modelsum.R
@@ -14,7 +14,7 @@ get_ms_part <- function(msList, modelnum, xList, yList, aList, sList, sValue, fa
     quasibinomial = cntrl$binomial.stats, binomial = cntrl$binomial.stats,
     quasipoisson = cntrl$poisson.stats, poisson = cntrl$poisson.stats,
     negbin = cntrl$negbin.stats, clog = cntrl$clog.stats, survival = cntrl$survival.stats,
-    ordinal = cntrl$ordinal.stats, gaussian = cntrl$gaussian.stats,
+    ordinal = cntrl$ordinal.stats, gaussian = cntrl$gaussian.stats, relrisk = cntrl$relrisk.stats
   )
 
   OUT <- NULL