+++ tests/testthat/test_modelsum.R
@@ -674,3 +674,19 @@ test_that("Nevents works (#266)", {
   )
 })
 
+test_that("relrisk works (#279)", {
+  opts <- options()
+  expect_identical(
+    capture.kable(summary(modelsum(mdquality.s ~ arm + sex, data = mockstudy, id = case, family = "relrisk"))),
+    c("|                            |RR    |CI.lower.RR |CI.upper.RR |p.value |Nmiss |",
+      "|:---------------------------|:-----|:-----------|:-----------|:-------|:-----|",
+      "|(Intercept)                 |0.890 |-0.152      |-0.081      |< 0.001 |252   |",
+      "|**Treatment Arm F: FOLFOX** |1.014 |-0.031      |0.059       |0.538   |      |",
+      "|**Treatment Arm G: IROX**   |1.021 |-0.029      |0.070       |0.412   |      |",
+      "|(Intercept)                 |0.899 |-0.130      |-0.083      |< 0.001 |252   |",
+      "|**sex Female**              |1.004 |-0.033      |0.042       |0.826   |      |"
+    )
+  )
+  options(opts)
+})
+